requirements.txt
# Django and DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-extensions==3.2.3
openai
# Database
psycopg2==2.9.10
pgvector==0.2.4

# Environment and configuration
python-decouple==3.8
django-environ==0.11.2

# API and HTTP
requests==2.31.0
httpx==0.25.2

# AI and ML
openai==1.3.7
numpy==1.24.3
scikit-learn==1.3.2

# Caching and performance
redis==5.0.1
django-redis==5.4.0
django-ratelimit==4.1.0

# Background tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Utilities
Pillow==10.1.0
python-slugify==8.0.1
python-dateutil==2.8.2
pytz==2023.3





# Documentation
drf-spectacular==0.26.5
drf-spectacular[sidecar]==0.26.5